fix(ci): clear quality gates for ledger validation
