+++ frontend/src/lib/components/EnforcementSettingsCardView.css
@@ -1,0 +1,208 @@
+.enforcement-policy-engine {
+	overflow: hidden;
+}
+
+.enforcement-policy-engine__header {
+	display: flex;
+	align-items: flex-start;
+	justify-content: space-between;
+	gap: 1rem;
+	margin-bottom: 1.25rem;
+}
+
+.enforcement-policy-engine__header h2 {
+	margin: 0.15rem 0 0;
+	color: var(--color-ink-50);
+	letter-spacing: 0;
+}
+
+.enforcement-policy-engine__header p:not(.enforcement-policy-engine__kicker) {
+	max-width: 44rem;
+	margin: 0.4rem 0 0;
+	color: var(--color-ink-400);
+	font-size: var(--text-xs);
+	line-height: 1.6;
+}
+
+.enforcement-policy-engine__kicker {
+	margin: 0;
+	color: var(--color-accent-300);
+	font-family: var(--font-mono);
+	font-size: 0.68rem;
+	font-weight: 800;
+	letter-spacing: 0;
+	text-transform: uppercase;
+}
+
+.enforcement-policy-engine__summary {
+	display: grid;
+	grid-template-columns: repeat(4, minmax(0, 1fr));
+	gap: 0.7rem;
+}
+
+.enforcement-policy-engine__summary-card,
+.enforcement-policy-engine__rail,
+.enforcement-policy-engine__callout {
+	border: 1px solid rgb(128 154 176 / 0.18);
+	border-radius: var(--radius-md);
+	background:
+		linear-gradient(180deg, rgb(24 32 40 / 0.74), rgb(10 13 18 / 0.56)), rgb(10 13 18 / 0.5);
+}
+
+.enforcement-policy-engine__summary-card {
+	display: grid;
+	gap: 0.35rem;
+	min-height: 6.1rem;
+	padding: 0.85rem;
+	color: var(--color-accent-300);
+}
+
+.enforcement-policy-engine__summary-card span,
+.enforcement-policy-engine__rail-state span {
+	color: var(--color-ink-400);
+	font-size: var(--text-xs);
+	font-weight: 700;
+	line-height: 1.25;
+}
+
+.enforcement-policy-engine__summary-card strong {
+	align-self: end;
+	color: var(--color-ink-50);
+	font-family: var(--font-mono);
+	font-size: clamp(0.96rem, 1.6vw, 1.25rem);
+	font-weight: 850;
+	line-height: 1.1;
+	overflow-wrap: anywhere;
+}
+
+.enforcement-policy-engine__callout {
+	display: flex;
+	align-items: flex-start;
+	gap: 0.65rem;
+	padding: 0.85rem 1rem;
+	color: var(--color-accent-300);
+	background: rgb(34 211 238 / 0.07);
+}
+
+.enforcement-policy-engine__callout p {
+	margin: 0;
+	color: var(--color-ink-300);
+	font-size: var(--text-sm);
+	line-height: 1.55;
+}
+
+.enforcement-policy-engine__rail-list {
+	display: grid;
+	gap: 0.7rem;
+}
+
+.enforcement-policy-engine__rail {
+	position: relative;
+	display: grid;
+	grid-template-columns: minmax(0, 1fr) minmax(8.5rem, auto);
+	align-items: center;
+	gap: 1rem;
+	padding: 0.9rem 1rem;
+}
+
+.enforcement-policy-engine__rail::before {
+	position: absolute;
+	inset: 0 auto 0 0;
+	width: 3px;
+	border-radius: var(--radius-full);
+	background: var(--color-accent-400);
+	content: '';
+}
+
+.enforcement-policy-engine__rail--hard::before {
+	background: var(--color-danger-400);
+}
+
+.enforcement-policy-engine__rail--soft::before {
+	background: var(--color-warning-400);
+}
+
+.enforcement-policy-engine__rail--shadow::before,
+.enforcement-policy-engine__rail--info::before {
+	background: var(--color-accent-400);
+}
+
+.enforcement-policy-engine__rail--enabled::before {
+	background: var(--color-success-400);
+}
+
+.enforcement-policy-engine__rail--disabled::before {
+	background: var(--color-ink-500);
+}
+
+.enforcement-policy-engine__rail-top {
+	display: flex;
+	flex-wrap: wrap;
+	align-items: center;
+	gap: 0.45rem;
+	color: var(--color-ink-50);
+	font-size: var(--text-sm);
+	font-weight: 780;
+}
+
+.enforcement-policy-engine__rail-top strong {
+	padding: 0.12rem 0.45rem;
+	border: 1px solid rgb(244 63 94 / 0.32);
+	border-radius: var(--radius-sm);
+	color: var(--color-danger-400);
+	background: rgb(244 63 94 / 0.1);
+	font-family: var(--font-mono);
+	font-size: 0.63rem;
+	font-weight: 800;
+	text-transform: uppercase;
+}
+
+.enforcement-policy-engine__rail p {
+	max-width: 46rem;
+	margin: 0.28rem 0 0;
+	color: var(--color-ink-400);
+	font-size: var(--text-xs);
+	line-height: 1.5;
+}
+
+.enforcement-policy-engine__rail-state {
+	display: grid;
+	gap: 0.25rem;
+	justify-items: end;
+	text-align: right;
+}
+
+.enforcement-policy-engine__rail-state strong {
+	color: var(--color-ink-50);
+	font-family: var(--font-mono);
+	font-size: var(--text-sm);
+	font-weight: 850;
+	overflow-wrap: anywhere;
+}
+
+@media (max-width: 900px) {
+	.enforcement-policy-engine__summary {
+		grid-template-columns: repeat(2, minmax(0, 1fr));
+	}
+}
+
+@media (max-width: 640px) {
+	.enforcement-policy-engine__header,
+	.enforcement-policy-engine__rail {
+		grid-template-columns: 1fr;
+		align-items: stretch;
+	}
+
+	.enforcement-policy-engine__header {
+		flex-direction: column;
+	}
+
+	.enforcement-policy-engine__summary {
+		grid-template-columns: 1fr;
+	}
+
+	.enforcement-policy-engine__rail-state {
+		justify-items: start;
+		text-align: left;
+	}
+}
