+++ reports/audit/staging.audit.report.json
@@ -19,10 +19,10 @@
   "incorrect_claims": [
     {
       "claim": "Testing has 5,358 tests",
-      "correction": "A fresh backend collection run reported 6469 tests collected.",
+      "correction": "A fresh backend collection run reported 6484 tests collected.",
       "evidence": [
         {
-          "line": 6469,
+          "line": 6484,
           "path": "pytest --collect-only"
         }
       ]
@@ -42,13 +42,13 @@
     "Live cutover evidence and operator sign-off are validated outside this repository."
   ],
   "measured_facts": {
-    "backend_tests_collected": 6469,
+    "backend_tests_collected": 6484,
     "direct_audit_logger_call_sites": 33,
     "fastapi_requirement": "~=0.128.0",
     "frontend_svelte_version": "5.53.0",
     "frontend_sveltekit_version": "2.57.1",
     "python_requires": ">=3.12,<3.13",
-    "test_and_spec_files": 965,
+    "test_and_spec_files": 968,
     "zombie_plugin_classes": 56
   },
   "partial_or_overstated_claims": [
@@ -66,14 +66,14 @@
   ],
   "product_name": "Valdrics",
   "repository": "CloudSentinel-AI",
-  "snapshot_date": "2026-06-03",
+  "snapshot_date": "2026-06-05",
   "source_report_label": "Repository-generated Codebase Audit Baseline",
   "summary": {
     "high_confidence_findings": [
       "The dashboard uses SvelteKit, Svelte 5, TypeScript, Tailwind CSS v4, Vitest, and Playwright."
     ],
     "important_corrections": [
-      "Backend tests collected currently total 6469, not 5358.",
+      "Backend tests collected currently total 6484, not 5358.",
       "Zombie detection plugin classes total 56 across providers, not 11."
     ],
     "overall": "Repository audit baseline refreshed from live measured facts."
