+++ tests/unit/ops/test_verify_new_frontend_disposition_register.py
@@ -12,7 +12,9 @@ def _write(path: Path, content: str) -> None:
 
 
 def _write_register(repo_root: Path, entries: list[dict]) -> Path:
-    register_path = repo_root / "docs/architecture/new_frontend_disposition_register.json"
+    register_path = (
+        repo_root / "docs/architecture/new_frontend_disposition_register.json"
+    )
     _write(
         register_path,
         json.dumps(
@@ -25,10 +27,15 @@ def _write_register(repo_root: Path, entries: list[dict]) -> Path:
     return register_path
 
 
-def test_verify_register_accepts_complete_pending_and_migrated_entries(tmp_path: Path) -> None:
+def test_verify_register_accepts_complete_pending_and_migrated_entries(
+    tmp_path: Path,
+) -> None:
     _write(tmp_path / "new_frontend/Sidebar.svelte", "<nav />")
     _write(tmp_path / "new_frontend/Future.svelte", "<section />")
-    _write(tmp_path / "frontend/src/routes/layout/AppAuthenticatedShell.svelte", "<aside />")
+    _write(
+        tmp_path / "frontend/src/routes/layout/AppAuthenticatedShell.svelte",
+        "<aside />",
+    )
     register_path = _write_register(
         tmp_path,
         [
@@ -36,7 +43,9 @@ def test_verify_register_accepts_complete_pending_and_migrated_entries(tmp_path:
                 "source_file": "new_frontend/Sidebar.svelte",
                 "status": "migrated",
                 "decision": "Migrated into the authenticated shell.",
-                "target_paths": ["frontend/src/routes/layout/AppAuthenticatedShell.svelte"],
+                "target_paths": [
+                    "frontend/src/routes/layout/AppAuthenticatedShell.svelte"
+                ],
                 "evidence": ["browser QA"],
                 "deletion_blocker": "",
             },
@@ -65,7 +74,9 @@ def test_verify_register_fails_when_new_frontend_file_is_missing_from_register(
     assert "entries must be a non-empty list" in errors
 
 
-def test_verify_register_fails_when_migrated_target_does_not_exist(tmp_path: Path) -> None:
+def test_verify_register_fails_when_migrated_target_does_not_exist(
+    tmp_path: Path,
+) -> None:
     _write(tmp_path / "new_frontend/Sidebar.svelte", "<nav />")
     register_path = _write_register(
         tmp_path,
@@ -83,10 +94,15 @@ def test_verify_register_fails_when_migrated_target_does_not_exist(tmp_path: Pat
 
     errors = verify_register(repo_root=tmp_path, register_path=register_path)
 
-    assert "new_frontend/Sidebar.svelte: target path does not exist: frontend/src/routes/layout/Missing.svelte" in errors
+    assert (
+        "new_frontend/Sidebar.svelte: target path does not exist: frontend/src/routes/layout/Missing.svelte"
+        in errors
+    )
 
 
-def test_verify_register_fails_when_pending_entry_has_no_blocker(tmp_path: Path) -> None:
+def test_verify_register_fails_when_pending_entry_has_no_blocker(
+    tmp_path: Path,
+) -> None:
     _write(tmp_path / "new_frontend/Future.svelte", "<section />")
     register_path = _write_register(
         tmp_path,
@@ -104,8 +120,39 @@ def test_verify_register_fails_when_pending_entry_has_no_blocker(tmp_path: Path)
 
     errors = verify_register(repo_root=tmp_path, register_path=register_path)
 
-    assert "new_frontend/Future.svelte: pending entries require deletion_blocker" in errors
+    assert (
+        "new_frontend/Future.svelte: pending entries require deletion_blocker" in errors
+    )
+
+
+def test_verify_register_accepts_archived_sources_when_handoff_root_is_absent(
+    tmp_path: Path,
+) -> None:
+    _write(
+        tmp_path / "frontend/src/routes/layout/AppAuthenticatedShell.svelte",
+        "<aside />",
+    )
+    register_path = _write_register(
+        tmp_path,
+        [
+            {
+                "source_file": "new_frontend/Sidebar.svelte",
+                "status": "migrated",
+                "decision": "Archived handoff source after migration.",
+                "target_paths": [
+                    "frontend/src/routes/layout/AppAuthenticatedShell.svelte"
+                ],
+                "evidence": ["browser QA"],
+                "deletion_blocker": "",
+            }
+        ],
+    )
+
+    assert verify_register(repo_root=tmp_path, register_path=register_path) == []
 
 
 def test_main_resolves_repo_relative_register_path() -> None:
-    assert main(["--register", "docs/architecture/new_frontend_disposition_register.json"]) == 0
+    assert (
+        main(["--register", "docs/architecture/new_frontend_disposition_register.json"])
+        == 0
+    )
